Sort member page view by the most active members
It's interesting for the user to see the most active members first

There are different ways to consider the most active member.

It can be the one with the most number of activities.
We preferred that because it is best for small size data.

But for a big dataset it might be better to consider members
 with most recent activities in a small timeline